+++ src/agent0/utils/__init__.py
@@ -1,7 +1,7 @@
 """General utility functions"""
 
 from .async_runner import async_runner
-from .block_before_timestamp import block_number_before_timestamp
+from .block_number_before_timestamp import block_number_before_timestamp
 
 __all__ = [
     "async_runner",
